improved protobufs, better request/response handling
