--- abci-protobuf/.gitignore
@@ -1,5 +1,0 @@
-.DS_Store
-/target/
-**/*.rs.bk
-Cargo.lock
-/.idea/
