--- abci-protobuf/README.md
@@ -1,11 +1,0 @@
-# ABCI Protobuf types
-
-Autogenerated protobuf files for the ABCI protocol.  
-
-See `version.txt` for the current supported version of Tendermint and ABCI
-
-## To Update
-1. Run `make update-proto`. This will grab the latest protobuf files from Tendermint master, and update the version.txt file.
-2. Run `cargo build`
-
-
