+++ build.rs
@@ -2,32 +2,15 @@ extern crate protobuf_codegen_pure;
 
 fn main() {
     protobuf_codegen_pure::run(protobuf_codegen_pure::Args {
-        out_dir: "src",
-        input: &["protos/github.com/tendermint/tendermint/libs/common/common.proto"],
-        includes: &["protos"],
+        out_dir: "src/messages",
+        input: &[
+            "protobuf/abci.proto",
+            "protobuf/github.com/tendermint/tendermint/libs/common/types.proto",
+            "protobuf/github.com/tendermint/tendermint/crypto/merkle/merkle.proto",
+        ],
+        includes: &["protobuf"],
         customize: protobuf_codegen_pure::Customize {
             ..Default::default()
         },
-    })
-    .expect("protoc");
-
-    protobuf_codegen_pure::run(protobuf_codegen_pure::Args {
-        out_dir: "src",
-        input: &["protos/github.com/tendermint/tendermint/crypto/merkle/merkle.proto"],
-        includes: &["protos"],
-        customize: protobuf_codegen_pure::Customize {
-            ..Default::default()
-        },
-    })
-    .expect("protoc");
-
-    protobuf_codegen_pure::run(protobuf_codegen_pure::Args {
-        out_dir: "src",
-        input: &["src/types.proto"],
-        includes: &["src", "protos"],
-        customize: protobuf_codegen_pure::Customize {
-            ..Default::default()
-        },
-    })
-    .expect("protoc");
+    }).expect("protoc");
 }
