+++ src/messages/abci.rs
@@ -11283,9 +11283,9 @@ static file_descriptor_proto_data: &'static [u8] = b"\
     \x12&\n\tvalidator\x18\x02\x20\x01(\x0b2\x0f.abci.ValidatorB\x02\x18\0\
     \x12\x12\n\x06height\x18\x03\x20\x01(\x03B\x02\x18\0\x12,\n\x04time\x18\
     \x04\x20\x01(\x0b2\x1a.google.protobuf.TimestampB\x02\x18\0\x12\x1e\n\
-    \x12total_voting_power\x18\x05\x20\x01(\x03B\x02\x18\0B\x1c\xd0\xe2\x1e\
-    \x01\xa8\xe2\x1e\x01\xb8\xe2\x1e\x01\xc8\xe2\x1e\x01\xe0\xe2\x1e\x01\xf8\
-    \xe1\x1e\x01\xc0\xe3\x1e\x01b\x06proto3\
+    \x12total_voting_power\x18\x05\x20\x01(\x03B\x02\x18\0B\x1c\xc0\xe3\x1e\
+    \x01\xb8\xe2\x1e\x01\xc8\xe2\x1e\x01\xd0\xe2\x1e\x01\xe0\xe2\x1e\x01\xa8\
+    \xe2\x1e\x01\xf8\xe1\x1e\x01b\x06proto3\
 ";
 
 static mut file_descriptor_proto_lazy: ::protobuf::lazy::Lazy<::protobuf::descriptor::FileDescriptorProto> = ::protobuf::lazy::Lazy {
