+++ src/messages/merkle.rs
@@ -22,21 +22,48 @@ use protobuf::Message as Message_imported_for_functions;
 use protobuf::ProtobufEnum as ProtobufEnum_imported_for_functions;
 
 #[derive(PartialEq,Clone,Default)]
-pub struct KVPair {
+pub struct ProofOp {
     // message fields
+    pub field_type: ::std::string::String,
     pub key: ::std::vec::Vec<u8>,
-    pub value: ::std::vec::Vec<u8>,
+    pub data: ::std::vec::Vec<u8>,
     // special fields
     pub unknown_fields: ::protobuf::UnknownFields,
     pub cached_size: ::protobuf::CachedSize,
 }
 
-impl KVPair {
-    pub fn new() -> KVPair {
+impl ProofOp {
+    pub fn new() -> ProofOp {
         ::std::default::Default::default()
     }
 
-    // bytes key = 1;
+    // string type = 1;
+
+    pub fn clear_field_type(&mut self) {
+        self.field_type.clear();
+    }
+
+    // Param is passed by value, moved
+    pub fn set_field_type(&mut self, v: ::std::string::String) {
+        self.field_type = v;
+    }
+
+    // Mutable pointer to the field.
+    // If field is not initialized, it is initialized with default value first.
+    pub fn mut_field_type(&mut self) -> &mut ::std::string::String {
+        &mut self.field_type
+    }
+
+    // Take field
+    pub fn take_field_type(&mut self) -> ::std::string::String {
+        ::std::mem::replace(&mut self.field_type, ::std::string::String::new())
+    }
+
+    pub fn get_field_type(&self) -> &str {
+        &self.field_type
+    }
+
+    // bytes key = 2;
 
     pub fn clear_key(&mut self) {
         self.key.clear();
@@ -62,34 +89,34 @@ impl KVPair {
         &self.key
     }
 
-    // bytes value = 2;
+    // bytes data = 3;
 
-    pub fn clear_value(&mut self) {
-        self.value.clear();
+    pub fn clear_data(&mut self) {
+        self.data.clear();
     }
 
     // Param is passed by value, moved
-    pub fn set_value(&mut self, v: ::std::vec::Vec<u8>) {
-        self.value = v;
+    pub fn set_data(&mut self, v: ::std::vec::Vec<u8>) {
+        self.data = v;
     }
 
     // Mutable pointer to the field.
     // If field is not initialized, it is initialized with default value first.
-    pub fn mut_value(&mut self) -> &mut ::std::vec::Vec<u8> {
-        &mut self.value
+    pub fn mut_data(&mut self) -> &mut ::std::vec::Vec<u8> {
+        &mut self.data
     }
 
     // Take field
-    pub fn take_value(&mut self) -> ::std::vec::Vec<u8> {
-        ::std::mem::replace(&mut self.value, ::std::vec::Vec::new())
+    pub fn take_data(&mut self) -> ::std::vec::Vec<u8> {
+        ::std::mem::replace(&mut self.data, ::std::vec::Vec::new())
     }
 
-    pub fn get_value(&self) -> &[u8] {
-        &self.value
+    pub fn get_data(&self) -> &[u8] {
+        &self.data
     }
 }
 
-impl ::protobuf::Message for KVPair {
+impl ::protobuf::Message for ProofOp {
     fn is_initialized(&self) -> bool {
         true
     }
@@ -99,10 +126,13 @@ impl ::protobuf::Message for KVPair {
             let (field_number, wire_type) = is.read_tag_unpack()?;
             match field_number {
                 1 => {
-                    ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.key)?;
+                    ::protobuf::rt::read_singular_proto3_string_into(wire_type, is, &mut self.field_type)?;
                 },
                 2 => {
-                    ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.value)?;
+                    ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.key)?;
+                },
+                3 => {
+                    ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.data)?;
                 },
                 _ => {
                     ::protobuf::rt::read_unknown_or_skip_group(field_number, wire_type, is, self.mut_unknown_fields())?;
@@ -116,11 +146,14 @@ impl ::protobuf::Message for KVPair {
     #[allow(unused_variables)]
     fn compute_size(&self) -> u32 {
         let mut my_size = 0;
-        if !self.key.is_empty() {
-            my_size += ::protobuf::rt::bytes_size(1, &self.key);
+        if !self.field_type.is_empty() {
+            my_size += ::protobuf::rt::string_size(1, &self.field_type);
         }
-        if !self.value.is_empty() {
-            my_size += ::protobuf::rt::bytes_size(2, &self.value);
+        if !self.key.is_empty() {
+            my_size += ::protobuf::rt::bytes_size(2, &self.key);
+        }
+        if !self.data.is_empty() {
+            my_size += ::protobuf::rt::bytes_size(3, &self.data);
         }
         my_size += ::protobuf::rt::unknown_fields_size(self.get_unknown_fields());
         self.cached_size.set(my_size);
@@ -128,11 +161,14 @@ impl ::protobuf::Message for KVPair {
     }
 
     fn write_to_with_cached_sizes(&self, os: &mut ::protobuf::CodedOutputStream) -> ::protobuf::ProtobufResult<()> {
-        if !self.key.is_empty() {
-            os.write_bytes(1, &self.key)?;
+        if !self.field_type.is_empty() {
+            os.write_string(1, &self.field_type)?;
         }
-        if !self.value.is_empty() {
-            os.write_bytes(2, &self.value)?;
+        if !self.key.is_empty() {
+            os.write_bytes(2, &self.key)?;
+        }
+        if !self.data.is_empty() {
+            os.write_bytes(3, &self.data)?;
         }
         os.write_unknown_fields(self.get_unknown_fields())?;
         ::std::result::Result::Ok(())
@@ -164,8 +200,8 @@ impl ::protobuf::Message for KVPair {
         Self::descriptor_static()
     }
 
-    fn new() -> KVPair {
-        KVPair::new()
+    fn new() -> ProofOp {
+        ProofOp::new()
     }
 
     fn descriptor_static() -> &'static ::protobuf::reflect::MessageDescriptor {
@@ -176,18 +212,23 @@ impl ::protobuf::Message for KVPair {
         unsafe {
             descriptor.get(|| {
                 let mut fields = ::std::vec::Vec::new();
+                fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeString>(
+                    "type",
+                    |m: &ProofOp| { &m.field_type },
+                    |m: &mut ProofOp| { &mut m.field_type },
+                ));
                 fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeBytes>(
                     "key",
-                    |m: &KVPair| { &m.key },
-                    |m: &mut KVPair| { &mut m.key },
+                    |m: &ProofOp| { &m.key },
+                    |m: &mut ProofOp| { &mut m.key },
                 ));
                 fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeBytes>(
-                    "value",
-                    |m: &KVPair| { &m.value },
-                    |m: &mut KVPair| { &mut m.value },
+                    "data",
+                    |m: &ProofOp| { &m.data },
+                    |m: &mut ProofOp| { &mut m.data },
                 ));
-                ::protobuf::reflect::MessageDescriptor::new::<KVPair>(
-                    "KVPair",
+                ::protobuf::reflect::MessageDescriptor::new::<ProofOp>(
+                    "ProofOp",
                     fields,
                     file_descriptor_proto()
                 )
@@ -195,96 +236,85 @@ impl ::protobuf::Message for KVPair {
         }
     }
 
-    fn default_instance() -> &'static KVPair {
-        static mut instance: ::protobuf::lazy::Lazy<KVPair> = ::protobuf::lazy::Lazy {
+    fn default_instance() -> &'static ProofOp {
+        static mut instance: ::protobuf::lazy::Lazy<ProofOp> = ::protobuf::lazy::Lazy {
             lock: ::protobuf::lazy::ONCE_INIT,
-            ptr: 0 as *const KVPair,
+            ptr: 0 as *const ProofOp,
         };
         unsafe {
-            instance.get(KVPair::new)
+            instance.get(ProofOp::new)
         }
     }
 }
 
-impl ::protobuf::Clear for KVPair {
+impl ::protobuf::Clear for ProofOp {
     fn clear(&mut self) {
+        self.clear_field_type();
         self.clear_key();
-        self.clear_value();
+        self.clear_data();
         self.unknown_fields.clear();
     }
 }
 
-impl ::std::fmt::Debug for KVPair {
+impl ::std::fmt::Debug for ProofOp {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
         ::protobuf::text_format::fmt(self, f)
     }
 }
 
-impl ::protobuf::reflect::ProtobufValue for KVPair {
+impl ::protobuf::reflect::ProtobufValue for ProofOp {
     fn as_ref(&self) -> ::protobuf::reflect::ProtobufValueRef {
         ::protobuf::reflect::ProtobufValueRef::Message(self)
     }
 }
 
 #[derive(PartialEq,Clone,Default)]
-pub struct KI64Pair {
+pub struct Proof {
     // message fields
-    pub key: ::std::vec::Vec<u8>,
-    pub value: i64,
+    pub ops: ::protobuf::RepeatedField<ProofOp>,
     // special fields
     pub unknown_fields: ::protobuf::UnknownFields,
     pub cached_size: ::protobuf::CachedSize,
 }
 
-impl KI64Pair {
-    pub fn new() -> KI64Pair {
+impl Proof {
+    pub fn new() -> Proof {
         ::std::default::Default::default()
     }
 
-    // bytes key = 1;
+    // repeated .merkle.ProofOp ops = 1;
 
-    pub fn clear_key(&mut self) {
-        self.key.clear();
+    pub fn clear_ops(&mut self) {
+        self.ops.clear();
     }
 
     // Param is passed by value, moved
-    pub fn set_key(&mut self, v: ::std::vec::Vec<u8>) {
-        self.key = v;
+    pub fn set_ops(&mut self, v: ::protobuf::RepeatedField<ProofOp>) {
+        self.ops = v;
     }
 
     // Mutable pointer to the field.
-    // If field is not initialized, it is initialized with default value first.
-    pub fn mut_key(&mut self) -> &mut ::std::vec::Vec<u8> {
-        &mut self.key
+    pub fn mut_ops(&mut self) -> &mut ::protobuf::RepeatedField<ProofOp> {
+        &mut self.ops
     }
 
     // Take field
-    pub fn take_key(&mut self) -> ::std::vec::Vec<u8> {
-        ::std::mem::replace(&mut self.key, ::std::vec::Vec::new())
+    pub fn take_ops(&mut self) -> ::protobuf::RepeatedField<ProofOp> {
+        ::std::mem::replace(&mut self.ops, ::protobuf::RepeatedField::new())
     }
 
-    pub fn get_key(&self) -> &[u8] {
-        &self.key
-    }
-
-    // int64 value = 2;
-
-    pub fn clear_value(&mut self) {
-        self.value = 0;
-    }
-
-    // Param is passed by value, moved
-    pub fn set_value(&mut self, v: i64) {
-        self.value = v;
-    }
-
-    pub fn get_value(&self) -> i64 {
-        self.value
+    pub fn get_ops(&self) -> &[ProofOp] {
+        &self.ops
     }
 }
 
-impl ::protobuf::Message for KI64Pair {
+impl ::protobuf::Message for Proof {
     fn is_initialized(&self) -> bool {
+        for v in &self.ops {
+            if !v.is_initialized() {
+                return false;
+            }
+        };
         true
     }
 
@@ -293,14 +323,7 @@ impl ::protobuf::Message for KI64Pair {
             let (field_number, wire_type) = is.read_tag_unpack()?;
             match field_number {
                 1 => {
-                    ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.key)?;
-                },
-                2 => {
-                    if wire_type != ::protobuf::wire_format::WireTypeVarint {
-                        return ::std::result::Result::Err(::protobuf::rt::unexpected_wire_type(wire_type));
-                    }
-                    let tmp = is.read_int64()?;
-                    self.value = tmp;
+                    ::protobuf::rt::read_repeated_message_into(wire_type, is, &mut self.ops)?;
                 },
                 _ => {
                     ::protobuf::rt::read_unknown_or_skip_group(field_number, wire_type, is, self.mut_unknown_fields())?;
@@ -314,24 +337,21 @@ impl ::protobuf::Message for KI64Pair {
     #[allow(unused_variables)]
     fn compute_size(&self) -> u32 {
         let mut my_size = 0;
-        if !self.key.is_empty() {
-            my_size += ::protobuf::rt::bytes_size(1, &self.key);
-        }
-        if self.value != 0 {
-            my_size += ::protobuf::rt::value_size(2, self.value, ::protobuf::wire_format::WireTypeVarint);
-        }
+        for value in &self.ops {
+            let len = value.compute_size();
+            my_size += 1 + ::protobuf::rt::compute_raw_varint32_size(len) + len;
+        };
         my_size += ::protobuf::rt::unknown_fields_size(self.get_unknown_fields());
         self.cached_size.set(my_size);
         my_size
     }
 
     fn write_to_with_cached_sizes(&self, os: &mut ::protobuf::CodedOutputStream) -> ::protobuf::ProtobufResult<()> {
-        if !self.key.is_empty() {
-            os.write_bytes(1, &self.key)?;
-        }
-        if self.value != 0 {
-            os.write_int64(2, self.value)?;
-        }
+        for v in &self.ops {
+            os.write_tag(1, ::protobuf::wire_format::WireTypeLengthDelimited)?;
+            os.write_raw_varint32(v.get_cached_size())?;
+            v.write_to_with_cached_sizes(os)?;
+        };
         os.write_unknown_fields(self.get_unknown_fields())?;
         ::std::result::Result::Ok(())
     }
@@ -362,8 +382,8 @@ impl ::protobuf::Message for KI64Pair {
         Self::descriptor_static()
     }
 
-    fn new() -> KI64Pair {
-        KI64Pair::new()
+    fn new() -> Proof {
+        Proof::new()
     }
 
     fn descriptor_static() -> &'static ::protobuf::reflect::MessageDescriptor {
@@ -374,18 +394,13 @@ impl ::protobuf::Message for KI64Pair {
         unsafe {
             descriptor.get(|| {
                 let mut fields = ::std::vec::Vec::new();
-                fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeBytes>(
-                    "key",
-                    |m: &KI64Pair| { &m.key },
-                    |m: &mut KI64Pair| { &mut m.key },
+                fields.push(::protobuf::reflect::accessor::make_repeated_field_accessor::<_, ::protobuf::types::ProtobufTypeMessage<ProofOp>>(
+                    "ops",
+                    |m: &Proof| { &m.ops },
+                    |m: &mut Proof| { &mut m.ops },
                 ));
-                fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeInt64>(
-                    "value",
-                    |m: &KI64Pair| { &m.value },
-                    |m: &mut KI64Pair| { &mut m.value },
-                ));
-                ::protobuf::reflect::MessageDescriptor::new::<KI64Pair>(
-                    "KI64Pair",
+                ::protobuf::reflect::MessageDescriptor::new::<Proof>(
+                    "Proof",
                     fields,
                     file_descriptor_proto()
                 )
@@ -393,45 +408,43 @@ impl ::protobuf::Message for KI64Pair {
         }
     }
 
-    fn default_instance() -> &'static KI64Pair {
-        static mut instance: ::protobuf::lazy::Lazy<KI64Pair> = ::protobuf::lazy::Lazy {
+    fn default_instance() -> &'static Proof {
+        static mut instance: ::protobuf::lazy::Lazy<Proof> = ::protobuf::lazy::Lazy {
             lock: ::protobuf::lazy::ONCE_INIT,
-            ptr: 0 as *const KI64Pair,
+            ptr: 0 as *const Proof,
         };
         unsafe {
-            instance.get(KI64Pair::new)
+            instance.get(Proof::new)
         }
     }
 }
 
-impl ::protobuf::Clear for KI64Pair {
+impl ::protobuf::Clear for Proof {
     fn clear(&mut self) {
-        self.clear_key();
-        self.clear_value();
+        self.clear_ops();
         self.unknown_fields.clear();
     }
 }
 
-impl ::std::fmt::Debug for KI64Pair {
+impl ::std::fmt::Debug for Proof {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
         ::protobuf::text_format::fmt(self, f)
     }
 }
 
-impl ::protobuf::reflect::ProtobufValue for KI64Pair {
+impl ::protobuf::reflect::ProtobufValue for Proof {
     fn as_ref(&self) -> ::protobuf::reflect::ProtobufValueRef {
         ::protobuf::reflect::ProtobufValueRef::Message(self)
     }
 }
 
 static file_descriptor_proto_data: &'static [u8] = b"\
-    \n8github.com/tendermint/tendermint/libs/common/types.proto\x12\x06commo\
-    n\",\n\x06KVPair\x12\x0f\n\x03key\x18\x01\x20\x01(\x0cB\x02\x18\0\x12\
-    \x11\n\x05value\x18\x02\x20\x01(\x0cB\x02\x18\0\".\n\x08KI64Pair\x12\x0f\
-    \n\x03key\x18\x01\x20\x01(\x0cB\x02\x18\0\x12\x11\n\x05value\x18\x02\x20\
-    \x01(\x03B\x02\x18\0B\x1c\xe0\xe2\x1e\x01\xc0\xe3\x1e\x01\xb8\xe2\x1e\
-    \x01\xc8\xe2\x1e\x01\xd0\xe2\x1e\x01\xf8\xe1\x1e\x01\xa8\xe2\x1e\x01b\
-    \x06proto3\
+    \n;github.com/tendermint/tendermint/crypto/merkle/merkle.proto\x12\x06me\
+    rkle\">\n\x07ProofOp\x12\x10\n\x04type\x18\x01\x20\x01(\tB\x02\x18\0\x12\
+    \x0f\n\x03key\x18\x02\x20\x01(\x0cB\x02\x18\0\x12\x10\n\x04data\x18\x03\
+    \x20\x01(\x0cB\x02\x18\0\")\n\x05Proof\x12\x20\n\x03ops\x18\x01\x20\x03(\
+    \x0b2\x0f.merkle.ProofOpB\x02\x18\0B\x14\xe0\xe2\x1e\x01\xf8\xe1\x1e\x01\
+    \xd0\xe2\x1e\x01\xa8\xe2\x1e\x01\xc8\xe2\x1e\x01b\x06proto3\
 ";
 
 static mut file_descriptor_proto_lazy: ::protobuf::lazy::Lazy<::protobuf::descriptor::FileDescriptorProto> = ::protobuf::lazy::Lazy {
