+++ src/messages/mod.rs
@@ -1,0 +1,3 @@
+pub mod abci; // Core types
+pub mod merkle;
+pub mod types; // Common (KV Pairs)
