+++ src/messages/types.rs
@@ -22,48 +22,21 @@ use protobuf::Message as Message_imported_for_functions;
 use protobuf::ProtobufEnum as ProtobufEnum_imported_for_functions;
 
 #[derive(PartialEq,Clone,Default)]
-pub struct ProofOp {
+pub struct KVPair {
     // message fields
-    pub field_type: ::std::string::String,
     pub key: ::std::vec::Vec<u8>,
-    pub data: ::std::vec::Vec<u8>,
+    pub value: ::std::vec::Vec<u8>,
     // special fields
     pub unknown_fields: ::protobuf::UnknownFields,
     pub cached_size: ::protobuf::CachedSize,
 }
 
-impl ProofOp {
-    pub fn new() -> ProofOp {
+impl KVPair {
+    pub fn new() -> KVPair {
         ::std::default::Default::default()
     }
 
-    // string type = 1;
-
-    pub fn clear_field_type(&mut self) {
-        self.field_type.clear();
-    }
-
-    // Param is passed by value, moved
-    pub fn set_field_type(&mut self, v: ::std::string::String) {
-        self.field_type = v;
-    }
-
-    // Mutable pointer to the field.
-    // If field is not initialized, it is initialized with default value first.
-    pub fn mut_field_type(&mut self) -> &mut ::std::string::String {
-        &mut self.field_type
-    }
-
-    // Take field
-    pub fn take_field_type(&mut self) -> ::std::string::String {
-        ::std::mem::replace(&mut self.field_type, ::std::string::String::new())
-    }
-
-    pub fn get_field_type(&self) -> &str {
-        &self.field_type
-    }
-
-    // bytes key = 2;
+    // bytes key = 1;
 
     pub fn clear_key(&mut self) {
         self.key.clear();
@@ -89,34 +62,34 @@ impl ProofOp {
         &self.key
     }
 
-    // bytes data = 3;
+    // bytes value = 2;
 
-    pub fn clear_data(&mut self) {
-        self.data.clear();
+    pub fn clear_value(&mut self) {
+        self.value.clear();
     }
 
     // Param is passed by value, moved
-    pub fn set_data(&mut self, v: ::std::vec::Vec<u8>) {
-        self.data = v;
+    pub fn set_value(&mut self, v: ::std::vec::Vec<u8>) {
+        self.value = v;
     }
 
     // Mutable pointer to the field.
     // If field is not initialized, it is initialized with default value first.
-    pub fn mut_data(&mut self) -> &mut ::std::vec::Vec<u8> {
-        &mut self.data
+    pub fn mut_value(&mut self) -> &mut ::std::vec::Vec<u8> {
+        &mut self.value
     }
 
     // Take field
-    pub fn take_data(&mut self) -> ::std::vec::Vec<u8> {
-        ::std::mem::replace(&mut self.data, ::std::vec::Vec::new())
+    pub fn take_value(&mut self) -> ::std::vec::Vec<u8> {
+        ::std::mem::replace(&mut self.value, ::std::vec::Vec::new())
     }
 
-    pub fn get_data(&self) -> &[u8] {
-        &self.data
+    pub fn get_value(&self) -> &[u8] {
+        &self.value
     }
 }
 
-impl ::protobuf::Message for ProofOp {
+impl ::protobuf::Message for KVPair {
     fn is_initialized(&self) -> bool {
         true
     }
@@ -126,13 +99,10 @@ impl ::protobuf::Message for ProofOp {
             let (field_number, wire_type) = is.read_tag_unpack()?;
             match field_number {
                 1 => {
-                    ::protobuf::rt::read_singular_proto3_string_into(wire_type, is, &mut self.field_type)?;
-                },
-                2 => {
                     ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.key)?;
                 },
-                3 => {
-                    ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.data)?;
+                2 => {
+                    ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.value)?;
                 },
                 _ => {
                     ::protobuf::rt::read_unknown_or_skip_group(field_number, wire_type, is, self.mut_unknown_fields())?;
@@ -146,14 +116,11 @@ impl ::protobuf::Message for ProofOp {
     #[allow(unused_variables)]
     fn compute_size(&self) -> u32 {
         let mut my_size = 0;
-        if !self.field_type.is_empty() {
-            my_size += ::protobuf::rt::string_size(1, &self.field_type);
-        }
         if !self.key.is_empty() {
-            my_size += ::protobuf::rt::bytes_size(2, &self.key);
+            my_size += ::protobuf::rt::bytes_size(1, &self.key);
         }
-        if !self.data.is_empty() {
-            my_size += ::protobuf::rt::bytes_size(3, &self.data);
+        if !self.value.is_empty() {
+            my_size += ::protobuf::rt::bytes_size(2, &self.value);
         }
         my_size += ::protobuf::rt::unknown_fields_size(self.get_unknown_fields());
         self.cached_size.set(my_size);
@@ -161,14 +128,11 @@ impl ::protobuf::Message for ProofOp {
     }
 
     fn write_to_with_cached_sizes(&self, os: &mut ::protobuf::CodedOutputStream) -> ::protobuf::ProtobufResult<()> {
-        if !self.field_type.is_empty() {
-            os.write_string(1, &self.field_type)?;
-        }
         if !self.key.is_empty() {
-            os.write_bytes(2, &self.key)?;
+            os.write_bytes(1, &self.key)?;
         }
-        if !self.data.is_empty() {
-            os.write_bytes(3, &self.data)?;
+        if !self.value.is_empty() {
+            os.write_bytes(2, &self.value)?;
         }
         os.write_unknown_fields(self.get_unknown_fields())?;
         ::std::result::Result::Ok(())
@@ -200,8 +164,8 @@ impl ::protobuf::Message for ProofOp {
         Self::descriptor_static()
     }
 
-    fn new() -> ProofOp {
-        ProofOp::new()
+    fn new() -> KVPair {
+        KVPair::new()
     }
 
     fn descriptor_static() -> &'static ::protobuf::reflect::MessageDescriptor {
@@ -212,23 +176,18 @@ impl ::protobuf::Message for ProofOp {
         unsafe {
             descriptor.get(|| {
                 let mut fields = ::std::vec::Vec::new();
-                fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeString>(
-                    "type",
-                    |m: &ProofOp| { &m.field_type },
-                    |m: &mut ProofOp| { &mut m.field_type },
-                ));
                 fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeBytes>(
                     "key",
-                    |m: &ProofOp| { &m.key },
-                    |m: &mut ProofOp| { &mut m.key },
+                    |m: &KVPair| { &m.key },
+                    |m: &mut KVPair| { &mut m.key },
                 ));
                 fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeBytes>(
-                    "data",
-                    |m: &ProofOp| { &m.data },
-                    |m: &mut ProofOp| { &mut m.data },
+                    "value",
+                    |m: &KVPair| { &m.value },
+                    |m: &mut KVPair| { &mut m.value },
                 ));
-                ::protobuf::reflect::MessageDescriptor::new::<ProofOp>(
-                    "ProofOp",
+                ::protobuf::reflect::MessageDescriptor::new::<KVPair>(
+                    "KVPair",
                     fields,
                     file_descriptor_proto()
                 )
@@ -236,85 +195,96 @@ impl ::protobuf::Message for ProofOp {
         }
     }
 
-    fn default_instance() -> &'static ProofOp {
-        static mut instance: ::protobuf::lazy::Lazy<ProofOp> = ::protobuf::lazy::Lazy {
+    fn default_instance() -> &'static KVPair {
+        static mut instance: ::protobuf::lazy::Lazy<KVPair> = ::protobuf::lazy::Lazy {
             lock: ::protobuf::lazy::ONCE_INIT,
-            ptr: 0 as *const ProofOp,
+            ptr: 0 as *const KVPair,
         };
         unsafe {
-            instance.get(ProofOp::new)
+            instance.get(KVPair::new)
         }
     }
 }
 
-impl ::protobuf::Clear for ProofOp {
+impl ::protobuf::Clear for KVPair {
     fn clear(&mut self) {
-        self.clear_field_type();
         self.clear_key();
-        self.clear_data();
+        self.clear_value();
         self.unknown_fields.clear();
     }
 }
 
-impl ::std::fmt::Debug for ProofOp {
+impl ::std::fmt::Debug for KVPair {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
         ::protobuf::text_format::fmt(self, f)
     }
 }
 
-impl ::protobuf::reflect::ProtobufValue for ProofOp {
+impl ::protobuf::reflect::ProtobufValue for KVPair {
     fn as_ref(&self) -> ::protobuf::reflect::ProtobufValueRef {
         ::protobuf::reflect::ProtobufValueRef::Message(self)
     }
 }
 
 #[derive(PartialEq,Clone,Default)]
-pub struct Proof {
+pub struct KI64Pair {
     // message fields
-    pub ops: ::protobuf::RepeatedField<ProofOp>,
+    pub key: ::std::vec::Vec<u8>,
+    pub value: i64,
     // special fields
     pub unknown_fields: ::protobuf::UnknownFields,
     pub cached_size: ::protobuf::CachedSize,
 }
 
-impl Proof {
-    pub fn new() -> Proof {
+impl KI64Pair {
+    pub fn new() -> KI64Pair {
         ::std::default::Default::default()
     }
 
-    // repeated .merkle.ProofOp ops = 1;
+    // bytes key = 1;
 
-    pub fn clear_ops(&mut self) {
-        self.ops.clear();
+    pub fn clear_key(&mut self) {
+        self.key.clear();
     }
 
     // Param is passed by value, moved
-    pub fn set_ops(&mut self, v: ::protobuf::RepeatedField<ProofOp>) {
-        self.ops = v;
+    pub fn set_key(&mut self, v: ::std::vec::Vec<u8>) {
+        self.key = v;
     }
 
     // Mutable pointer to the field.
-    pub fn mut_ops(&mut self) -> &mut ::protobuf::RepeatedField<ProofOp> {
-        &mut self.ops
+    // If field is not initialized, it is initialized with default value first.
+    pub fn mut_key(&mut self) -> &mut ::std::vec::Vec<u8> {
+        &mut self.key
     }
 
     // Take field
-    pub fn take_ops(&mut self) -> ::protobuf::RepeatedField<ProofOp> {
-        ::std::mem::replace(&mut self.ops, ::protobuf::RepeatedField::new())
+    pub fn take_key(&mut self) -> ::std::vec::Vec<u8> {
+        ::std::mem::replace(&mut self.key, ::std::vec::Vec::new())
     }
 
-    pub fn get_ops(&self) -> &[ProofOp] {
-        &self.ops
+    pub fn get_key(&self) -> &[u8] {
+        &self.key
+    }
+
+    // int64 value = 2;
+
+    pub fn clear_value(&mut self) {
+        self.value = 0;
+    }
+
+    // Param is passed by value, moved
+    pub fn set_value(&mut self, v: i64) {
+        self.value = v;
+    }
+
+    pub fn get_value(&self) -> i64 {
+        self.value
     }
 }
 
-impl ::protobuf::Message for Proof {
+impl ::protobuf::Message for KI64Pair {
     fn is_initialized(&self) -> bool {
-        for v in &self.ops {
-            if !v.is_initialized() {
-                return false;
-            }
-        };
         true
     }
 
@@ -323,7 +293,14 @@ impl ::protobuf::Message for Proof {
             let (field_number, wire_type) = is.read_tag_unpack()?;
             match field_number {
                 1 => {
-                    ::protobuf::rt::read_repeated_message_into(wire_type, is, &mut self.ops)?;
+                    ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.key)?;
+                },
+                2 => {
+                    if wire_type != ::protobuf::wire_format::WireTypeVarint {
+                        return ::std::result::Result::Err(::protobuf::rt::unexpected_wire_type(wire_type));
+                    }
+                    let tmp = is.read_int64()?;
+                    self.value = tmp;
                 },
                 _ => {
                     ::protobuf::rt::read_unknown_or_skip_group(field_number, wire_type, is, self.mut_unknown_fields())?;
@@ -337,21 +314,24 @@ impl ::protobuf::Message for Proof {
     #[allow(unused_variables)]
     fn compute_size(&self) -> u32 {
         let mut my_size = 0;
-        for value in &self.ops {
-            let len = value.compute_size();
-            my_size += 1 + ::protobuf::rt::compute_raw_varint32_size(len) + len;
-        };
+        if !self.key.is_empty() {
+            my_size += ::protobuf::rt::bytes_size(1, &self.key);
+        }
+        if self.value != 0 {
+            my_size += ::protobuf::rt::value_size(2, self.value, ::protobuf::wire_format::WireTypeVarint);
+        }
         my_size += ::protobuf::rt::unknown_fields_size(self.get_unknown_fields());
         self.cached_size.set(my_size);
         my_size
     }
 
     fn write_to_with_cached_sizes(&self, os: &mut ::protobuf::CodedOutputStream) -> ::protobuf::ProtobufResult<()> {
-        for v in &self.ops {
-            os.write_tag(1, ::protobuf::wire_format::WireTypeLengthDelimited)?;
-            os.write_raw_varint32(v.get_cached_size())?;
-            v.write_to_with_cached_sizes(os)?;
-        };
+        if !self.key.is_empty() {
+            os.write_bytes(1, &self.key)?;
+        }
+        if self.value != 0 {
+            os.write_int64(2, self.value)?;
+        }
         os.write_unknown_fields(self.get_unknown_fields())?;
         ::std::result::Result::Ok(())
     }
@@ -382,8 +362,8 @@ impl ::protobuf::Message for Proof {
         Self::descriptor_static()
     }
 
-    fn new() -> Proof {
-        Proof::new()
+    fn new() -> KI64Pair {
+        KI64Pair::new()
     }
 
     fn descriptor_static() -> &'static ::protobuf::reflect::MessageDescriptor {
@@ -394,13 +374,18 @@ impl ::protobuf::Message for Proof {
         unsafe {
             descriptor.get(|| {
                 let mut fields = ::std::vec::Vec::new();
-                fields.push(::protobuf::reflect::accessor::make_repeated_field_accessor::<_, ::protobuf::types::ProtobufTypeMessage<ProofOp>>(
-                    "ops",
-                    |m: &Proof| { &m.ops },
-                    |m: &mut Proof| { &mut m.ops },
+                fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeBytes>(
+                    "key",
+                    |m: &KI64Pair| { &m.key },
+                    |m: &mut KI64Pair| { &mut m.key },
                 ));
-                ::protobuf::reflect::MessageDescriptor::new::<Proof>(
-                    "Proof",
+                fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeInt64>(
+                    "value",
+                    |m: &KI64Pair| { &m.value },
+                    |m: &mut KI64Pair| { &mut m.value },
+                ));
+                ::protobuf::reflect::MessageDescriptor::new::<KI64Pair>(
+                    "KI64Pair",
                     fields,
                     file_descriptor_proto()
                 )
@@ -408,43 +393,45 @@ impl ::protobuf::Message for Proof {
         }
     }
 
-    fn default_instance() -> &'static Proof {
-        static mut instance: ::protobuf::lazy::Lazy<Proof> = ::protobuf::lazy::Lazy {
+    fn default_instance() -> &'static KI64Pair {
+        static mut instance: ::protobuf::lazy::Lazy<KI64Pair> = ::protobuf::lazy::Lazy {
             lock: ::protobuf::lazy::ONCE_INIT,
-            ptr: 0 as *const Proof,
+            ptr: 0 as *const KI64Pair,
         };
         unsafe {
-            instance.get(Proof::new)
+            instance.get(KI64Pair::new)
         }
     }
 }
 
-impl ::protobuf::Clear for Proof {
+impl ::protobuf::Clear for KI64Pair {
     fn clear(&mut self) {
-        self.clear_ops();
+        self.clear_key();
+        self.clear_value();
         self.unknown_fields.clear();
     }
 }
 
-impl ::std::fmt::Debug for Proof {
+impl ::std::fmt::Debug for KI64Pair {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
         ::protobuf::text_format::fmt(self, f)
     }
 }
 
-impl ::protobuf::reflect::ProtobufValue for Proof {
+impl ::protobuf::reflect::ProtobufValue for KI64Pair {
     fn as_ref(&self) -> ::protobuf::reflect::ProtobufValueRef {
         ::protobuf::reflect::ProtobufValueRef::Message(self)
     }
 }
 
 static file_descriptor_proto_data: &'static [u8] = b"\
-    \n;github.com/tendermint/tendermint/crypto/merkle/merkle.proto\x12\x06me\
-    rkle\">\n\x07ProofOp\x12\x10\n\x04type\x18\x01\x20\x01(\tB\x02\x18\0\x12\
-    \x0f\n\x03key\x18\x02\x20\x01(\x0cB\x02\x18\0\x12\x10\n\x04data\x18\x03\
-    \x20\x01(\x0cB\x02\x18\0\")\n\x05Proof\x12\x20\n\x03ops\x18\x01\x20\x03(\
-    \x0b2\x0f.merkle.ProofOpB\x02\x18\0B\x14\xd0\xe2\x1e\x01\xe0\xe2\x1e\x01\
-    \xa8\xe2\x1e\x01\xc8\xe2\x1e\x01\xf8\xe1\x1e\x01b\x06proto3\
+    \n8github.com/tendermint/tendermint/libs/common/types.proto\x12\x06commo\
+    n\",\n\x06KVPair\x12\x0f\n\x03key\x18\x01\x20\x01(\x0cB\x02\x18\0\x12\
+    \x11\n\x05value\x18\x02\x20\x01(\x0cB\x02\x18\0\".\n\x08KI64Pair\x12\x0f\
+    \n\x03key\x18\x01\x20\x01(\x0cB\x02\x18\0\x12\x11\n\x05value\x18\x02\x20\
+    \x01(\x03B\x02\x18\0B\x1c\xe0\xe2\x1e\x01\xc8\xe2\x1e\x01\xc0\xe3\x1e\
+    \x01\xd0\xe2\x1e\x01\xa8\xe2\x1e\x01\xf8\xe1\x1e\x01\xb8\xe2\x1e\x01b\
+    \x06proto3\
 ";
 
 static mut file_descriptor_proto_lazy: ::protobuf::lazy::Lazy<::protobuf::descriptor::FileDescriptorProto> = ::protobuf::lazy::Lazy {
